# Performance improvements for Polygon validation