--- a/geom/heap.go
+++ b/geom/heap.go
@@ -1,5 +1,13 @@
 package geom
 
+func seq(n int) []int {
+	s := make([]int, n)
+	for i := range s {
+		s[i] = i
+	}
+	return s
+}
+
 type intHeap struct {
 	less func(i, j int) bool
 	data []int
@@ -10,7 +18,7 @@ func (h *intHeap) push(elem int) {
 	i := len(h.data) - 1
 	for i > 0 {
 		parent := (i - 1) / 2
-		if h.less(parent, i) {
+		if h.lt(parent, i) {
 			break
 		}
 		h.data[parent], h.data[i] = h.data[i], h.data[parent]
@@ -28,22 +36,22 @@ func (h *intHeap) pop() {
 		childB := 2*i + 2
 		switch {
 		case childA < len(h.data) && childB < len(h.data):
-			if h.less(i, childA) {
-				if h.less(childB, i) {
+			if h.lt(i, childA) {
+				if h.lt(childB, i) {
 					swapWith = childB
 				}
 			} else {
 				swapWith = childA
-				if h.less(childB, childA) {
+				if h.lt(childB, childA) {
 					swapWith = childB
 				}
 			}
 		case childA < len(h.data):
-			if h.less(childA, i) {
+			if h.lt(childA, i) {
 				swapWith = childA
 			}
 		case childB < len(h.data):
-			if h.less(childB, i) {
+			if h.lt(childB, i) {
 				swapWith = childB
 			}
 		}
@@ -53,4 +61,8 @@ func (h *intHeap) pop() {
 		h.data[swapWith], h.data[i] = h.data[i], h.data[swapWith]
 		i = swapWith
 	}
+}
+
+func (h *intHeap) lt(i, j int) bool {
+	return h.less(h.data[i], h.data[j])
 }