… geom/line_heap.go
… geom/line_heap_test.go