Primarily these are micro-optimisations relating to memory usage. There is a small regression for MLS x MLS hasIntersection, most pronounced at small input sizes. That small regression seems to be worth it when compared to the improvements in other cases.

```
benchmark                                               old ns/op     new ns/op     delta
BenchmarkIntersectsLineStringWithLineString/n=10        2187          2495          +14.08%
BenchmarkIntersectsLineStringWithLineString/n=100       22053         23813         +7.98%
BenchmarkIntersectsLineStringWithLineString/n=1000      277166        287880        +3.87%
BenchmarkIntersectsLineStringWithLineString/n=10000     3086197       3146948       +1.97%
BenchmarkPolygonSingleRingValidation/n=10               8221          2996          -63.56%
BenchmarkPolygonSingleRingValidation/n=100              83416         34277         -58.91%
BenchmarkPolygonSingleRingValidation/n=1000             914922        421470        -53.93%
BenchmarkPolygonSingleRingValidation/n=10000            10411253      5059155       -51.41%
BenchmarkPolygonMultipleRingsValidation/n=4             51496         18740         -63.61%
BenchmarkPolygonMultipleRingsValidation/n=36            790795        323139        -59.14%
BenchmarkPolygonMultipleRingsValidation/n=400           22310689      10686561      -52.10%
BenchmarkPolygonMultipleRingsValidation/n=4096          829429151     349774385     -57.83%

benchmark                                               old allocs     new allocs     delta
BenchmarkIntersectsLineStringWithLineString/n=10        17             21             +23.53%
BenchmarkIntersectsLineStringWithLineString/n=100       17             21             +23.53%
BenchmarkIntersectsLineStringWithLineString/n=1000      17             21             +23.53%
BenchmarkIntersectsLineStringWithLineString/n=10000     17             21             +23.53%
BenchmarkPolygonSingleRingValidation/n=10               104            18             -82.69%
BenchmarkPolygonSingleRingValidation/n=100              905            21             -97.68%
BenchmarkPolygonSingleRingValidation/n=1000             8912           24             -99.73%
BenchmarkPolygonSingleRingValidation/n=10000            89709          34             -99.96%
BenchmarkPolygonMultipleRingsValidation/n=4             797            169            -78.80%
BenchmarkPolygonMultipleRingsValidation/n=36            12338          3096           -74.91%
BenchmarkPolygonMultipleRingsValidation/n=400           352232         114274         -67.56%
BenchmarkPolygonMultipleRingsValidation/n=4096          10545142       3463224        -67.16%

benchmark                                               old bytes     new bytes     delta
BenchmarkIntersectsLineStringWithLineString/n=10        1152          1216          +5.56%
BenchmarkIntersectsLineStringWithLineString/n=100       6976          7040          +0.92%
BenchmarkIntersectsLineStringWithLineString/n=1000      66112         66176         +0.10%
BenchmarkIntersectsLineStringWithLineString/n=10000     655936        656000        +0.01%
BenchmarkPolygonSingleRingValidation/n=10               4526          1352          -70.13%
BenchmarkPolygonSingleRingValidation/n=100              39858         9336          -76.58%
BenchmarkPolygonSingleRingValidation/n=1000             379498        73976         -80.51%
BenchmarkPolygonSingleRingValidation/n=10000            4910886       1843448       -62.46%
BenchmarkPolygonMultipleRingsValidation/n=4             26930         7072          -73.74%
BenchmarkPolygonMultipleRingsValidation/n=36            398482        126496        -68.26%
BenchmarkPolygonMultipleRingsValidation/n=400           11134437      4552130       -59.12%
BenchmarkPolygonMultipleRingsValidation/n=4096          330618944     136951370     -58.58%
```